feat(nc-android): add runtime permission request
